NeaImage and NeaSpectrum modules replaced by images, readers, spectra modules
