+++ pySNOM/__init__.py
@@ -3,11 +3,15 @@ Module pySNOM
 =============
 Scanning Near-field Optical Microscopy (SNOM) analysis tools
 """
-
-from .NeaImager import NeaImage
-from .NeaSpectra import NeaSpectrum
+from .defaults import defaults
+from .images import Image, Process
+from .readers import Reader
+from .spectra import NeaSpectrum, NeaInterferogram
+# import images
+# import spectra
+# import readers
 
 __version__ = '0.0.2'
 __author__ = 'Gergely Németh, Ferenc Borondics'
 __credits__ = 'Wigner Research Centre for Physics, Synchrotron SOLEIL'
-__all__ = ["NeaImage", "NeaSpectrum"]
+__all__ = ["defaults","Reader","Image","Process","NeaSpectrum","NeaInterferogram"]
