+++ pySNOM/tests/test_NeaImage.py
@@ -4,13 +4,15 @@ import os
 import numpy as np
 
 import pySNOM
-from pySNOM import NeaImage
+from pySNOM import Reader, Image
 
 class test_NeaImage(unittest.TestCase):
     def test_readfile(self):
         f = 'datasets/image.gwy'
-        im = NeaImage()
-        im.read_from_gwyfile(os.path.join(pySNOM.__path__[0], f), 'M1A raw')
+        file_reader = Reader(os.path.join(pySNOM.__path__[0], f))
+        channeldata = file_reader.read_gwychannel('M1A raw')
+
+        im = Image(channeldata=channeldata)
 
         np.testing.assert_almost_equal(im.data[7][7], 70.70956420)
         np.testing.assert_almost_equal(im.data[8][8], 70.34189605)
