+++ requirements.txt
@@ -1,4 +1,4 @@
 gwyfile>=0.2.0
-numpy>=2.0.0
-setuptools>=70.0.0
-scipy>=1.12.0
+numpy
+setuptools
+scipy
